+++ assets/fonts/font-awesome.php
@@ -10,6 +10,6 @@ if (!defined('ABSPATH')) {
 }
 ?>
 
-<link href="<?php echo LOOPIS_HQ_THEME_URI; ?>/assets/fonts/css/fontawesome.min.css" rel="stylesheet" />
-<link href="<?php echo LOOPIS_HQ_THEME_URI; ?>/assets/fonts/css/regular.min.css" rel="stylesheet" />
-<link href="<?php echo LOOPIS_HQ_THEME_URI; ?>/assets/fonts/css/solid.min.css" rel="stylesheet" />
+<link href="<?php echo LOOPIS_THEME_HQ_URI; ?>/assets/fonts/css/fontawesome.min.css" rel="stylesheet" />
+<link href="<?php echo LOOPIS_THEME_HQ_URI; ?>/assets/fonts/css/regular.min.css" rel="stylesheet" />
+<link href="<?php echo LOOPIS_THEME_HQ_URI; ?>/assets/fonts/css/solid.min.css" rel="stylesheet" />