+++ front-page.php
@@ -8,16 +8,14 @@ get_header(); ?>
 
 <div class="content">
     <div class="page-padding">
-<h1>Ge & få saker på ett nytt sätt</h1>
+<h5>Det nya sättet att ge & få saker</h5>
 <hr>
-<p>Paxa med en knapptryckning, hämta i skåpet.<br>
-LOOPIS gör det enkelt & roligt att ge & få saker i ditt grannskap.</p>
-
-<?php include_once LOOPIS_HQ_THEME_DIR . '/templates/access/message.php'; ?>
+<p>Paxa i appen, hämta i skåpet. Enkelt och roligt!</p>
+<p><span class="big-link"><a href="/faq/hur-funkar-LOOPIS">📌 Hur funkar LOOPIS?</a></span></p>
 
 <div class="frontpage-map">
             <div class="frontpage-map__legend">
-                <h5>📍 Områden</h5>
+                <h5>🗺 Karta</h5>
                 <hr>
                 <p>❤ Här finns LOOPIS</p>
                 <p>🧡 Här öppnar snart LOOPIS</p>
@@ -29,7 +27,7 @@ LOOPIS gör det enkelt & roligt att ge & få saker i ditt grannskap.</p>
         <div class="frontpage-interest-form">
             <h2>Vill du ha LOOPIS i ditt område?</h2>
             <hr>
-            <p>Fyll i formuläret! När 100 personer med ditt postnummer har anmält intresse kan vi börja planera för att öppna LOOPIS.</p>
+            <p>Vi behöver 100 personer med ditt postnummer för att börja planera.</p>
             <form class="interest-form" method="post" action="" novalidate>
                 <?php wp_nonce_field( 'loopis_interest_form', 'loopis_interest_nonce' ); ?>
                 <div class="interest-form__field">
@@ -44,7 +42,7 @@ LOOPIS gör det enkelt & roligt att ge & få saker i ditt grannskap.</p>
             </form>
         </div><!-- frontpage-interest-form -->
 
-        <?php if (!is_user_logged_in()) include_once LOOPIS_HQ_THEME_DIR . '/templates/access/visitor.php'; ?>
+        <?php if (!is_user_logged_in()) include_once LOOPIS_THEME_HQ_DIR . '/templates/access/visitor.php'; ?>
 
     </div><!--page-padding-->
 </div><!--content-->