+++ header.php
@@ -10,15 +10,15 @@
 	if (is_home() || is_front_page()) {
 		$title = $blog_name . '.app';
 		$description = 'Ge & få saker av dina grannar.';
-		$meta_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_app.png';
-		$og_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_og.png';
+		$meta_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_app.png';
+		$og_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_og.png';
 		$og_type = 'website';
 		$og_url = home_url('/');
 	} elseif (is_single() || is_page()) {
         $title = get_the_title() . ' - ' . $blog_name;
         $description = get_the_excerpt();
-        $meta_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_app.png';
-        $og_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_og.png';
+        $meta_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_app.png';
+        $og_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_og.png';
         if (has_post_thumbnail()) {
             $thumbnail = wp_get_attachment_image_src(get_post_thumbnail_id($post->ID), 'large');
             if ($thumbnail) {
@@ -32,36 +32,36 @@
 		$author = get_queried_object();
 		$title = esc_html($author->display_name) . ' - ' . $blog_name;
 		$description = esc_html($author->description);
-		$meta_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_app.png';
-		$og_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_og.png';
+		$meta_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_app.png';
+		$og_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_og.png';
 		$og_type = 'profile';
 		$og_url = get_author_posts_url($author->ID);
 	} elseif (is_tag()) {
 		$title = '#' . single_tag_title('', false) . ' - ' . $blog_name;
 		$description = 'Kategorier på LOOPIS.app';
-		$meta_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_app.png';
-		$og_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_og.png';
+		$meta_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_app.png';
+		$og_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_og.png';
 		$og_type = 'website';
 		$og_url = get_tag_link(get_queried_object_id());
 	} elseif (is_archive()) {
         $title = $blog_name;
         $description = 'För en glad och hållbar framtid.';
-        $meta_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_app.png';
-        $og_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_og.png';
+        $meta_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_app.png';
+        $og_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_og.png';
         $og_type = 'website';
         $og_url = home_url('/');
     } elseif (is_404()) {
         $title = 'Hoppsan! - ' . $blog_name;
         $description = 'Sidan kunde inte hittas.';
-        $meta_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_app.png';
-        $og_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_og.png';
+        $meta_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_app.png';
+        $og_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_og.png';
         $og_type = 'website';
         $og_url = home_url('/');
     } else {
         $title = $blog_name;
         $description = 'För en glad och hållbar framtid.';
-        $meta_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_app.png';
-        $og_image = LOOPIS_HQ_THEME_URI . '/assets/img/LOOPIS_og.png';
+        $meta_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_app.png';
+        $og_image = LOOPIS_THEME_HQ_URI . '/assets/img/LOOPIS_og.png';
         $og_type = 'website';
         $og_url = home_url('/');
     }
@@ -77,8 +77,8 @@
 	<?php if (is_single()) { echo '<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=1">'; }
 	else { echo '<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=0">'; } ?>
 	<!--Fonts-->
-	<?php include_once LOOPIS_HQ_THEME_DIR . '/assets/fonts/google-fonts.php'; ?>
-	<?php include_once LOOPIS_HQ_THEME_DIR . '/assets/fonts/font-awesome.php'; ?>
+	<?php include_once LOOPIS_THEME_HQ_DIR . '/assets/fonts/google-fonts.php'; ?>
+	<?php include_once LOOPIS_THEME_HQ_DIR . '/assets/fonts/font-awesome.php'; ?>
 	<!--Favicon-->
 	<link rel="canonical" href="<?php echo esc_url($og_url); ?>">
 	<link rel="manifest" href="/favicon/site.webmanifest">
@@ -112,7 +112,7 @@
 	<header id="header">
 		<div class="group">
 			<div class="header-back" onclick="history.back()"><i class="fas fa-chevron-left"></i></div>
-			<a href="<?php echo esc_url(home_url('/')); ?>"><img src="<?php echo LOOPIS_HQ_THEME_URI; ?>/assets/img/LOOPIS_logo_fat.png" alt="LOOPIS-logo" id="header-img"></a>
+			<a href="<?php echo esc_url(home_url('/')); ?>"><img src="<?php echo LOOPIS_THEME_HQ_URI; ?>/assets/img/LOOPIS_logo.png" alt="LOOPIS-logo" id="header-img"></a>
 			<?php if ( is_front_page() && current_user_can('loopis_admin') && !current_user_can('administrator') ) : ?>
 				<div class="header-faq" onclick="location.href='<?php echo esc_url( home_url('/admin/') ); ?>'">🐙</div>
 			<?php else: ?>