+++ single-faq.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Template for single FAQ post.
+ * 
+ * Copy from local theme FAQ post layout.
+ */
+
+get_header(); ?>
+
+<div class="content">
+<div class="page-padding">
+
+<?php the_content(); ?>
+<div class="clear"></div>
+
+<!-- More questions? -->
+<?php include LOOPIS_THEME_HQ_DIR . '/templates/faq/questions-faq.php'; ?>
+
+</div><!--page-padding-->
+</div><!--content-->
+
+<?php get_footer(); ?>