+++ templates/faq/questions-visitor.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Block for routing visitor to FAQ on post
+ *
+ * Included in single.php
+ */
+
+if (!defined('ABSPATH')) {
+    exit; // Exit if accessed directly
+}   
+
+$faq_url = home_url( '/faq/' );
+?>
+
+<div class="columns"><div class="column1"><h5>💡 Vanliga frågor</h5></div>
+<div class="column2 bottom"><a href="<?php echo esc_url($faq_url);?>">Alla frågor & svar →</a></div></div>
+<hr>
+<p><span class="big-link"><a href="<?php echo esc_url($faq_url . 'hur-funkar-loopis/');?>">📌 Hur funkar LOOPIS?</a></span></p>
+<p><span class="big-link"><a href="<?php echo esc_url($faq_url . 'varfor-medlemskap/');?>">📌 Varför måste jag vara medlem?</a></span></p>
+<p><span class="big-link"><a href="<?php echo esc_url($faq_url . 'varfor-bagis/');?>">📌 Varför måste jag bo i Bagis?</a></span></p>