+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# 📋 Changelog for "LOOPIS HQ Theme"
+# 📋 Changelog for "LOOPIS Theme HQ"
 
-## 0.0 (2026-04-20)
+## 0.02 (2026-05-13)
+- Renamed ('loopis-hq-theme' > 'loopis-theme-hq') for better sorting of files
+- Preliminary templates for custom post types 'faq' and 'news'
+
+## 0.01 (2026-04-20)
 - Init commit with basic structure